Create index.html
This is a HTML file that contains a video calling application. Here is the explanation of the code:

The first few lines of code include meta tags and links to external style sheets and JavaScript files that are required for the webpage to function correctly. It includes Bootstrap CSS, Google Fonts, and a CSS file named videochat2.css.

The script section contains a function named "record" that uses the Web Speech API to convert speech to text in real-time when a user speaks into a text field with the ID of "speechToText". This enables live captioning generated through speech recognition for the video calling webapp.

The body of the HTML file contains a navbar with the title "Video Calling", a container with a button to start a new meeting, and a table with a text field to capture speech input.

There is also a div with the ID of "jitsi-container" that is empty initially but will be populated with the video calling interface once a user clicks on the "Start a new meeting" button.

The final part of the code includes optional JavaScript files such as jQuery, Popper.js, and Bootstrap JS that are required for certain functionalities to work correctly.